backend/requirements.txt
# FastAPI Backend Requirements

# FastAPI Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
email-validator==2.2.0

# LangChain and AI
langchain==0.3.27
langgraph==0.6.1
langchain-groq==0.2.1
langchain-community==0.3.27

# Search and APIs
tavily-python==0.5.0
wikipedia==1.4.0
googlemaps==4.10.0

# PDF Generation
reportlab==4.2.5
markdown-pdf==1.3.0
weasyprint==63.1

# Utilities
requests==2.31.0
python-dotenv==1.0.1
typing-extensions==4.12.2

# Email Service
resend==0.8.0

# SMS & Phone Service
twilio==9.0.4

# State Management
redis==5.0.1
hiredis==2.3.2

# Enhanced LangGraph
langgraph-checkpoint-sqlite>=2.0.1

# Utilities for Confirmation System
qrcode[pil]==7.4.2
jinja2==3.1.3
pillow==10.4.0
